pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "aleam"
dynamic = ["version"]
description = "True randomness for AI and machine learning. Non-recursive, stateless, cryptographically secure random number generator."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Fardin Sabid", email = "contact.fardinsabid@gmail.com"}
]
maintainers = [
    {name = "Fardin Sabid", email = "contact.fardinsabid@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security :: Cryptography",
]

dependencies = [
    "numpy>=2.4.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.0",
    "pytest-benchmark>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "ruff>=0.1",
    "pre-commit>=3.0",
    "build>=1.0",
    "twine>=4.0",
]
torch = [
    "torch>=2.0.0",
]
tensorflow = [
    "tensorflow>=2.12.0",
]
jax = [
    "jax>=0.4.0",
    "jaxlib>=0.4.0",
]
cupy = [
    "cupy-cuda12x>=12.0.0",
]
pandas = [
    "pandas>=1.5.0",
]
polars = [
    "polars>=0.18.0",
]
xarray = [
    "xarray>=2023.0.0",
]
pymc = [
    "pymc>=5.0.0",
]
scipy = [
    "scipy>=1.9.0",
]
all = [
    "torch>=2.0.0",
    "tensorflow>=2.12.0",
    "jax>=0.4.0",
    "jaxlib>=0.4.0",
    "cupy-cuda12x>=12.0.0",
    "pandas>=1.5.0",
    "polars>=0.18.0",
    "xarray>=2023.0.0",
    "pymc>=5.0.0",
    "scipy>=1.9.0",
]

[project.urls]
Homepage = "https://github.com/fardinsabid/aleam"
Repository = "https://github.com/fardinsabid/aleam"
Documentation = "https://github.com/fardinsabid/aleam/docs"
Issues = "https://github.com/fardinsabid/aleam/issues"
Changelog = "https://github.com/fardinsabid/aleam/releases"

[project.scripts]
aleam = "aleam.__main__:main"

[tool.setuptools]
packages = {find = {include = ["aleam*"]}}
include-package-data = true

[tool.setuptools.package-data]
aleam = [
    "py.typed",
    "*.pyi",
    "cuda_kernels.cu",
]

[tool.setuptools_scm]
write_to = "aleam/_version.py"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py38"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".git",
    ".venv",
    "build",
    "dist",
    "__pycache__",
]

[tool.ruff.isort]
known-first-party = ["aleam"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "build",
    "dist",
    "tests",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=aleam",
    "--cov-report=term-missing",
    "--cov-report=html",
    "-p", "no:cacheprovider",
]
markers = [
    "slow: marks tests as slow",
    "gpu: marks tests as requiring GPU",
    "integration: marks tests as integration tests",
]

[tool.pytest-benchmark]
group_by = "group"
warmup = true
warmup_iterations = 10

[tool.coverage.run]
source = ["aleam"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/_version.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
known_first_party = ["aleam"]

[tool.bumpversion]
current_version = "1.0.0"
commit = true
tag = true
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)(\\.(?P<release>[a-z]+)(?P<dev>\\d+))?"
serialize = [
    "{major}.{minor}.{patch}.{release}{dev}",
    "{major}.{minor}.{patch}",
]

[tool.bumpversion.parts]
release = "{release}"

[tool.bumpversion.files]
"pyproject.toml" = [
    {match = "current_version = \"(?P<version>.*?)\"", replace = "current_version = \"{new_version}\""},
]
"aleam/__init__.py" = [
    {match = "__version__ = \"(?P<version>.*?)\"", replace = "__version__ = \"{new_version}\""},
]
"setup.py" = [
    {match = "version=\"(?P<version>.*?)\"", replace = "version=\"{new_version}\""},
]

[tool.cibuildwheel]
build = "cp38-* cp39-* cp310-* cp311-* cp312-*"
skip = "pp* *-musllinux_*"
manylinux-x86_64-image = "manylinux2014"
test-command = "pytest {project}/tests"
test-requires = "pytest"
before-test = "pip install numpy"

[tool.cibuildwheel.linux]
before-build = "yum install -y openblas-devel"

[tool.cibuildwheel.macos]
before-build = "brew install openblas"

[tool.cibuildwheel.windows]
before-build = "choco install openblas"

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
fail-under = 80
exclude = ["tests", "docs", "build", "dist", "examples", "benchmarks"]